requirements.txt
python-telegram-bot==13.15
requests>=2.31.0
python-dotenv==0.21.1
pytz>=2024.1
